// src/constants/practice-area-enum.ts
export enum PracticeAreaEnum {
  FAMILY_LAW = "FAMILY_LAW",
  CRIMINAL_LAW = "CRIMINAL_LAW",
  COMMERCIAL_LAW = "COMMERCIAL_LAW",
  IMMIGRATION_LAW = "IMMIGRATION_LAW",
  PERSONAL_INJURY = "PERSONAL_INJURY",
  REAL_ESTATE = "REAL_ESTATE",
  LABOR_LAW = "LABOR_LAW",
  BANKRUPTCY = "BANKRUPTCY",
  TAX_LAW = "TAX_LAW",
  INTELLECTUAL_PROPERTY = "INTELLECTUAL_PROPERTY",
  INHERITANCE_LAW = "INHERITANCE_LAW",
  ENVIRONMENTAL_LAW = "ENVIRONMENTAL_LAW",
  MEDICAL_LIABILITY = "MEDICAL_LIABILITY",
  CONTRACT_LAW = "CONTRACT_LAW",
  CIVIL_RIGHTS = "CIVIL_RIGHTS",
}